refactor: ajuste de importacao pos stories
