+++ src/app/(main)/transacoes/page.tsx
@@ -11,20 +11,20 @@ import { Button } from "@/components/ui/button/button";
 import { Calendar } from "@/components/ui/calendar/calendar";
 import { Card } from "@/components/ui/card/card";
 import { DataTable } from "@/components/ui/dataTable/dataTable";
-import { Input } from "@/components/ui/input/input";
+import { Input } from "@/components/ui/inputs/input";
 import { Label } from "@/components/ui/label/label";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "@/components/ui/popover";
+} from "@/components/ui/popover/popover";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select";
+} from "@/components/ui/select/select";
 
 import { api } from "@/lib/api";
 import type {
