+++ src/components/auth/LoginForm.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/navigation";
 import Link from "next/link";
 import { authApi } from "@/lib/auth-api";
 import { Button } from "@/components/ui/button/button";
-import { Input } from "@/components/ui/input/input";
+import { Input } from "@/components/ui/inputs/input";
 import { Label } from "@/components/ui/label/label";
 import { PasswordInput } from "@/components/auth/PasswordInput";
