+++ src/components/auth/PasswordInput.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { Eye, EyeOff } from "lucide-react";
-import { Input } from "@/components/ui/input/input";
+import { Input } from "@/components/ui/inputs/input";
 
 interface PasswordInputProps {
   id: string;
