+++ src/components/transactions/components/TransactionModal.tsx
@@ -12,22 +12,22 @@ import {
   DialogClose,
 } from "@/components/ui/dialogModal/dialog";
 import { Button } from "@/components/ui/button/button";
-import { Input } from "@/components/ui/input/input";
+import { Input } from "@/components/ui/inputs/input";
 import { Label } from "@/components/ui/label/label";
 import { Calendar } from "@/components/ui/calendar/calendar";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "@/components/ui/popover";
+} from "@/components/ui/popover/popover";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select";
-import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
+} from "@/components/ui/select/select";
+import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs/tabs";
 import { TransactionFormType, TransactionModalProps } from "@/components/transactions/types";
 import { useTransactionForm } from "@/components/transactions/hooks/useTransactionForm";
