+++ src/components/ui/dataTable/dataTable.tsx
@@ -16,7 +16,7 @@ import {
   PaginationEllipsis,
   PaginationItem,
   PaginationLink
-} from "@/components/ui/pagination"
+} from "@/components/ui/pagination/pagination"
 
 import {
   Table,
@@ -25,7 +25,7 @@ import {
   TableHead,
   TableHeader,
   TableRow
-} from "@/components/ui/table"
+} from "@/components/ui/table/table"
 import { useState } from "react"
